WIP: Turn model into a proper Zipper
This roughly follows a Zipper pattern of a Rose Tree without the need to
move laterally and with static deep levels:
`Queue:Sequence:Step:Execution:Action`.

Also Queue is now a polymorphic container so that the same operations
can be performed for pending Queue of Actions or a Queue of results.

*Unconsing* is implemented but *consing* still needs to be implemented.
It should be fairly symmetrical to *unconsing*.