Reduce the number of generated `Arbitrary` `SequenceQueue`s
After adding the `Condition`s to `SequenceQueue`s the number generated data
samples with `Arbitrary` explodes.

I tried to combine the `Conditions` generator with `Sequence` list generator
with `Gen.zip` but it still does the product.

I also tried making the Sequences for the generator a Zip List but it seems the
there is no Applicative instance for `Gen` and I can't use a Zip List as a
Monad.

So in the end I just reduced the maximum size of the generated Sequence List.